new lineofsight algorithm implemented

+++ src/game/ingame/CoordinateObject.java
@@ -1,20 +1,20 @@
 package game.ingame;
 
 public abstract class CoordinateObject {
-	private int x, y;
+	private double x, y;
 
-	public void setX(int x) {
+	public void setX(double x) {
 		this.x = x;
 	}
 
-	public void setY(int y) {
+	public void setY(double y) {
 		this.y = y;
 	}
 	
-	public int getX() {
+	public double getX() {
 		return x;
 	}
-	public int getY() {
+	public double getY() {
 		return y;
 	}
 }
